Ensure identical build numbers for each parallel build
Separate version update into two scripts.

gitVersionFile.ts will run once before the parallel builds and provide version info

updateVersion.ts runs with with agent to update project files with version/build numbers
